Day9: Done!

+++ AdventOfCode-CSharp/Day7/Instruction.cs
@@ -7,18 +7,18 @@ namespace AdventOfCode_CSharp.Day7
 {
     public class Instruction
     {
-        public Instruction(List<int> instructionList)
+        public Instruction(List<long> instructionList)
         {
             Parse(instructionList);
         }
 
         public int Opcode { get; set; }
-        public Tuple<ParameterMode, int> Param1Value { get; set; }
-        public Tuple<ParameterMode, int> Param2Value { get; set; }
-        public Tuple<ParameterMode, int> Param3Value { get; set; }
+        public Tuple<ParameterMode, long> Param1Value { get; set; }
+        public Tuple<ParameterMode, long> Param2Value { get; set; }
+        public Tuple<ParameterMode, long> Param3Value { get; set; }
 
         private List<int> ParsedOpcodes { get; set; }
-        public void Parse(List<int> instructionList)
+        public void Parse(List<long> instructionList)
         {
             while(instructionList.Count < 4)
             {
@@ -34,22 +34,22 @@ namespace AdventOfCode_CSharp.Day7
             size -= 2;
 
             if (size > 0)
-                Param1Value = new Tuple<ParameterMode, int>((ParameterMode)ParsedOpcodes[--size], instructionList[1]);
+                Param1Value = new Tuple<ParameterMode, long>((ParameterMode)ParsedOpcodes[--size], instructionList[1]);
             else
-                Param1Value = new Tuple<ParameterMode, int>(0, instructionList[1]);
+                Param1Value = new Tuple<ParameterMode, long>(0, instructionList[1]);
 
             if (size > 0)
-                Param2Value = new Tuple<ParameterMode, int>((ParameterMode)ParsedOpcodes[--size], instructionList[2]);
+                Param2Value = new Tuple<ParameterMode, long>((ParameterMode)ParsedOpcodes[--size], instructionList[2]);
             else
-                Param2Value = new Tuple<ParameterMode, int>(0, instructionList[2]);
+                Param2Value = new Tuple<ParameterMode, long>(0, instructionList[2]);
 
             if (size > 0)
-                Param3Value = new Tuple<ParameterMode, int>((ParameterMode)ParsedOpcodes[--size], instructionList[3]);
+                Param3Value = new Tuple<ParameterMode, long>((ParameterMode)ParsedOpcodes[--size], instructionList[3]);
             else
-                Param3Value = new Tuple<ParameterMode, int>(0, instructionList[3]);
+                Param3Value = new Tuple<ParameterMode, long>(0, instructionList[3]);
         }
 
-        private void parseOpcode(int opcode)
+        private void parseOpcode(long opcode)
         {
             ParsedOpcodes = opcode.ToString().Select(i => int.Parse(i.ToString())).ToList();
